+++ demo/wxgame/native-audio-config.js
@@ -1,2 +1,0 @@
-GameGlobal.__godotMinigameNativeAudioMinDurationSeconds = 3;
-GameGlobal.__godotMinigameNativeAudioCacheLimitBytes = 8388608;