+++ templates/4.7/minigame4.7/engine/game.js
@@ -1,7 +1,0 @@
-
-import './native-audio-manifest'
-import './godot-sdk'
-import './godot'
-const exe = '/engine/godot';
-const pack = '/engine/empty-tips.bin';
-GODOTSDK.startGame(exe, pack)